requirements.txt
Flask==2.3.3
Werkzeug==2.3.8
Jinja2==3.1.6
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
requests==2.33.0
pytest==7.4.3
pytest-cov==4.1.0
esprima==4.0.1
Flask-Migrate==4.0.5
PyMySQL==1.1.1
pyotp>=2.9.0
qrcode>=7.4
Pillow>=10.0.0
